Update LICENSE
The AGPLv3 addresses the "software as a service" (SaaS) scenario. It is essentially the GPLv3 with an added provision that specifically deals with network use.
It requires that if you modify AGPLv3-licensed software and run it on a server, making it accessible to users over a network, you must provide the source code of the modified version to those users.
In essence, the AGPLv3 closes the "loophole" that exists in the GPLv3 regarding network services. It ensures that if you modify the software and provide it as a network service, you must also provide the corresponding source code.

Signed-off-by: Tatius-Wolff <85422568+Tatius-Wolff@users.noreply.github.com>